feat: splashscreen
Co-authored by: (zt) Nick <nwowens32@gmail.com>
