+++ app/build.gradle.kts
@@ -96,6 +96,8 @@ dependencies {
     implementation("androidx.compose.ui:ui:$composeVersion")
     implementation("androidx.compose.ui:ui-tooling-preview:$composeVersion")
     implementation("androidx.navigation:navigation-compose:2.5.0")
+    implementation("androidx.appcompat:appcompat:1.4.2")
+    implementation("androidx.core:core-splashscreen:1.0.0-rc01")
 
     // ReVanced
     implementation("app.revanced:revanced-patcher:2.4.0")
