+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <permission android:name="android.permission.QUERY_ALL_PACKAGES"
+    <permission
+        android:name="android.permission.QUERY_ALL_PACKAGES"
         tools:ignore="ReservedSystemPermission" />
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -22,17 +23,17 @@
 
     <application
         android:allowBackup="true"
+        android:extractNativeLibs="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:requestLegacyExternalStorage="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:requestLegacyExternalStorage="true"
-        android:extractNativeLibs="true"
-
         android:theme="@style/Theme.ReVancedManager"
+
         tools:targetApi="32">
         <activity
-            android:name=".ShellHeater"
+            android:name=".MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ReVancedManager">
 
@@ -41,11 +42,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity
-            android:name=".MainActivity"
-            android:exported="false"
-            android:theme="@style/Theme.ReVancedManager" />
     </application>
 
 </manifest>
