--- app/src/main/java/app/revanced/manager/ShellHeater.kt
@@ -1,29 +1,0 @@
-package app.revanced.manager
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import com.topjohnwu.superuser.Shell
-
-class ShellHeater : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        Shell.getShell {
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-    }
-
-    companion object {
-        init {
-            Shell.enableVerboseLogging = BuildConfig.DEBUG
-            Shell.setDefaultBuilder(
-                Shell.Builder.create()
-                    .setFlags(Shell.FLAG_REDIRECT_STDERR)
-                    .setTimeout(10)
-            )
-        }
-    }
-}
